# functions/src/controllers/swagger/bnb.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Binance Smart Chain API",
    "version": "1.0.0",
    "description": "API for handling BSC wallet operations, including wallet creation, balance retrieval, and transaction processing."
  },
  "paths": {
...
    "/wallet/balance/get": {
      "get": {
        "summary": "Retrieve wallet balance",
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "balance": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tx/bnb/create": {
      "post": {
        "summary": "Send a BNB or USDT transaction",
        "operationId": "sendTransaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "string"
                  },
                  "privateKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "from",
                  "to",
                  "amount",
                  "privateKey"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "txHash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/tx/usdt/create": {
      "post": {
        "summary": "Send BUSDT Transaction",
        "description": "Transfers Binance-Pegged USDT (BUSDT) to a recipient address.",
        "operationId": "sendBUSDT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Amount of BUSDT to send"
                  },
                  "to": {
                    "type": "string",
                    "format": "address",
                    "description": "Recipient's BSC address"
                  },
                  "privateKey": {
                    "type": "string",
                    "description": "Sender's private key"
                  }
                },
                "required": [
                  "amount",
                  "to",
                  "privateKey"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "txHash": {
                      "type": "string",
                      "description": "Transaction hash"
                    },
                    "fee": {
                      "type": "string",
                      "description": "Estimated transaction fee in ETH"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Transaction failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}